+++ src/indieweb_utils/utils/__init__.py
@@ -1,0 +1,3 @@
+from .urls import canonicalize_url
+
+__all__ = ["canonicalize_url"]
